Restructure things a bit and add css style to improve the sticky nav appearance
The message div is moved out above the user/problem nav in the main template.
The nav div is added to the gateway template.

This div that is now in both templates is styled with a box shadow to
set it off from the rest of the page.  This gives it a nav header
appearance that works well when sticky takes effect.

The padding that bootstrap 2.3.2 adds to the body on narrow screens is
overriden. This is something that version 2.3.2 of bootstrap does wrong.
Eating up the width with padding on an already narrow screen is just not
well though out.  Newer versions of bootstrap don't do this.  This makes
it so that the nav buttons don't wrap as often.  Long student names will
still make the student nav wrap though.